Made a decomposition
